refactor(providers): move cross-cutting utilities into providers/util.py
Moved Progress, _find_tsv_columns, _is_md5_checksum, read_checksum_file,
compute_md5, validate_download, _remove_if_exists, _get_download_url,
_resolve_local_path from Files into providers/util.py. Files keeps shim
re-exports so existing references keep working.

No behaviour change. Test suite green.
